BuildPhylogenies cpus:memory test
Even though all processes run successfully according to the Nextflow report, there are java memory error files (e.g. hs_err_pid*) in the working directory for this process. According to the error file the output file may therefore be truncated/incomplete (because out of MEmory error). In the nextflow report this process uses > 1000% of allowed CPUs and resubmitted for another attempt with double the initial memory.